move test fixture to conftest.py
