+++ aws/BUILD
@@ -1,0 +1,3 @@
+python_test_utils(
+    name="test_utils",
+)
