+++ aws/tests/test_manage_items.py
@@ -1,22 +1,7 @@
-from dataclasses import dataclass
 from unittest.mock import patch
-
-import pytest
 
 from aws.example_rest_api.api.manage_items import lambda_handler
 from aws.example_rest_api.models.item import Item
-
-
-@pytest.fixture
-def lambda_context():
-	@dataclass
-	class LambdaContext:
-		function_name: str = 'test'
-		memory_limit_in_mb: int = 128
-		invoked_function_arn: str = 'arn:aws:lambda:eu-west-1:809313241:function:test'
-		aws_request_id: str = '52fdfc07-2182-154f-163f-5f0f9a621d72'
-
-	return LambdaContext()
 
 
 @patch(
